Various small changes have been made:
Global Attributes TimeSeries and Profile is now using ver 2.1 of the ArcticNet Metadata formatting. Ready for public dispersion.
Nortek devices perform the in-air offset without pausing and there is now a single line comment. This is my preference, but it might not be for others.
Note: when running the QC filters do not use the Vertical Spike filter (this has not been made for Profile data only and causes an error when trying to perform QC filtering.
Some new scripts have been made for the ISUS nitrate sensor, though it doesn`t work yet.